requirements.txt
datetime
pyyaml
certifi
charset-normalizer
elastic-transport
elasticsearch
idna
pip
polars
pytz
requests
setuptools
urllib3
zope.interface
autocommand
backports.tarfile
importlib-metadata
inflect
jaraco.collections
jaraco.context
jaraco.functools
jaraco.text
more-itertools
packaging
platformdirs
tomli
typeguard
typing-extensions
wheel
zipp